requirements.txt
asgiref==3.11.0
attrs==25.4.0
dj-database-url==3.0.1
Django==4.2.27
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.15.3
django-stubs==4.2.7
drf-spectacular==0.29.0
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mypy==1.11.2
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.2
PyYAML==6.0.3
types-PyYAML==6.0.12.20241230
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.4
typing_extensions==4.15.0
uritemplate==4.2.0
whitenoise==6.11.0
